backend/requirements.txt
openstacksdk
fastapi[standard]
uvicorn
pydantic
python-dotenv